# frontend-react/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BillGen API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/invoices/preview": {
      "post": {
        "tags": [
          "invoices"
        ],
        "summary": "Preview",
        "description": "Authoritative totals while composing. No DB write, no sequence consumed.",
        "operationId": "preview_invoices_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvoicePreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoicePreviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/invoices": {
      "post": {
        "tags": [
          "invoices"
        ],
        "summary": "Create Invoice",
        "description": "Create a DRAFT invoice \u2014 no number is consumed. Finalize it with\nPOST /invoices/{id}/issue.\n\nThe monthly invoice allowance is consumed **here**, at creation, not at\nissue. Issuing is the legally load-bearing act and must never be the step\nthat fails for a commercial reason: a Belgian sole trader who has drafted an\ninvoice has to be able to finalize it.",
        "operationId": "create_invoice_invoices_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvoiceCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "invoices"
        ],
        "summary": "List Invoices",
        "operationId": "list_invoices_invoices_get",
        "parameters": [
          {
            "name": "company_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Company Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/InvoiceStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "client_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Client Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InvoiceResponse"
                  },
                  "title": "Response List Invoices Invoices Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/invoices/{invoice_id}/issue": {
      "post": {
        "tags": [
          "invoices"
        ],
        "summary": "Issue Invoice",
        "description": "Issue a draft: consume the gapless number, freeze totals, set ISSUED.",
        "operationId": "issue_invoice_invoices__invoice_id__issue_post",
        "parameters": [
          {
            "name": "invoice_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Invoice Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/IssueRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/invoices/{invoice_id}/duplicate": {
      "post": {
        "tags": [
          "invoices"
        ],
        "summary": "Duplicate Invoice",
        "description": "Copy an invoice into a new DRAFT dated today. No number is consumed and\nnothing on the source changes \u2014 including an issued or voided source, which\nis the common case (re-billing last month's work).",
        "operationId": "duplicate_invoice_invoices__invoice_id__duplicate_post",
        "parameters": [
          {
            "name": "invoice_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Invoice Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/invoices/{invoice_id}": {
      "delete": {
        "tags": [
          "invoices"
        ],
        "summary": "Delete Invoice",
        "description": "Hard-delete a DRAFT invoice. Issued invoices return 409 (never deleted).",
        "operationId": "delete_invoice_invoices__invoice_id__delete",
        "parameters": [
          {
            "name": "invoice_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Invoice Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "invoices"
        ],
        "summary": "Get Invoice",
        "operationId": "get_invoice_invoices__invoice_id__get",
        "parameters": [
          {
            "name": "invoice_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Invoice Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/invoices/{invoice_id}/void": {
      "post": {
        "tags": [
          "invoices"
        ],
        "summary": "Void Invoice",
        "operationId": "void_invoice_invoices__invoice_id__void_post",
        "parameters": [
          {
            "name": "invoice_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Invoice Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoidRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/invoices/{invoice_id}/html": {
      "get": {
        "tags": [
          "invoices"
        ],
        "summary": "Invoice Html",
        "operationId": "invoice_html_invoices__invoice_id__html_get",
        "parameters": [
          {
            "name": "invoice_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Invoice Id"
            }
          },
          {
            "name": "template",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Template"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/invoices/{invoice_id}/pdf": {
      "get": {
        "tags": [
          "invoices"
        ],
        "summary": "Invoice Pdf",
        "operationId": "invoice_pdf_invoices__invoice_id__pdf_get",
        "parameters": [
          {
            "name": "invoice_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Invoice Id"
            }
          },
          {
            "name": "template",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Template"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/invoices/{invoice_id}/peppol.xml": {
      "get": {
        "tags": [
          "invoices"
        ],
        "summary": "Invoice Peppol",
        "description": "Peppol BIS 3.0 XML. Every tier can do this \u2014 Peppol is the Belgian\ndifferentiator, not an upsell \u2014 but the monthly document allowance applies,\ncounted from the EXPORT_PEPPOL audit entries this call writes.",
        "operationId": "invoice_peppol_invoices__invoice_id__peppol_xml_get",
        "parameters": [
          {
            "name": "invoice_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Invoice Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/reports/invoices": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Invoice Report",
        "description": "Counts and totals per effective status, plus a monthly series. Replaces\naggregating the full invoice list in the browser. Omit `period` for all time.",
        "operationId": "invoice_report_reports_invoices_get",
        "parameters": [
          {
            "name": "company_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Company Id"
            }
          },
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^\\d{4}(?:-(?:Q[1-4]|0[1-9]|1[0-2]))?$"
                },
                {
                  "type": "null"
                }
              ],
              "examples": [
                "2026-Q3"
              ],
              "title": "Period"
            }
          },
          {
            "name": "today",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Today"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceReportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InvoiceCreateRequest": {
        "properties": {
          "company_id": {
            "type": "string",
            "format": "uuid",
            "title": "Company Id"
          },
          "client_id": {
            "type": "string",
            "format": "uuid",
            "title": "Client Id"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/InvoiceLineIn"
            },
            "type": "array",
            "minItems": 1,
            "title": "Lines"
          },
          "issue_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issue Date"
          },
          "due_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date"
          },
          "invoice_discount": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DiscountIn"
              },
              {
                "type": "null"
              }
            ]
          },
          "comments": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comments"
          },
          "payment_terms": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payment Terms"
          },
          "pdf_template": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pdf Template"
          },
          "currency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currency"
          }
        },
        "type": "object",
        "required": [
          "company_id",
          "client_id",
          "lines"
        ],
        "title": "InvoiceCreateRequest"
      },
...
      "InvoicePreviewRequest": {
        "properties": {
          "lines": {
            "items": {
              "$ref": "#/components/schemas/InvoiceLineIn"
            },
            "type": "array",
            "minItems": 1,
            "title": "Lines"
          },
          "invoice_discount": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DiscountIn"
              },
              {
                "type": "null"
              }
            ]
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "default": "EUR"
          }
        },
        "type": "object",
        "required": [
          "lines"
        ],
        "title": "InvoicePreviewRequest"
      },
      "InvoicePreviewResponse": {
        "properties": {
          "subtotal_ht": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Subtotal Ht"
          },
          "total_discount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Discount"
          },
          "net_ht": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Net Ht"
          },
          "total_vat": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Vat"
          },
          "total_ttc": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Ttc"
          },
          "vat_breakdown": {
            "additionalProperties": {
              "type": "string",
              "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
            },
            "type": "object",
            "title": "Vat Breakdown"
          }
        },
        "type": "object",
        "required": [
          "subtotal_ht",
          "total_discount",
          "net_ht",
          "total_vat",
          "total_ttc",
          "vat_breakdown"
        ],
        "title": "InvoicePreviewResponse"
      },
...
      "InvoiceResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "company_id": {
            "type": "string",
            "format": "uuid",
            "title": "Company Id"
          },
          "client_id": {
            "type": "string",
            "format": "uuid",
            "title": "Client Id"
          },
          "reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference"
          },
          "sequence_global": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sequence Global"
          },
          "issue_date": {
            "type": "string",
            "format": "date",
            "title": "Issue Date"
          },
          "due_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/InvoiceLineOut"
            },
            "type": "array",
            "title": "Lines"
          },
          "invoice_discount": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DiscountOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "comments": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comments"
          },
          "payment_terms": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payment Terms"
          },
          "pdf_template": {
            "type": "string",
            "title": "Pdf Template"
          },
          "subtotal_ht": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Subtotal Ht"
          },
          "total_discount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Discount"
          },
          "total_vat": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Vat"
          },
          "total_ttc": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Ttc"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "voided_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Voided At"
          },
          "voided_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Voided Reason"
          },
          "voided_by_credit_note_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Voided By Credit Note Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "company_id",
          "client_id",
          "reference",
          "sequence_global",
          "issue_date",
          "due_date",
          "currency",
          "lines",
          "invoice_discount",
          "comments",
          "payment_terms",
          "pdf_template",
          "subtotal_ht",
          "total_discount",
          "total_vat",
          "total_ttc",
          "status",
          "voided_at",
          "voided_reason",
          "voided_by_credit_note_id"
        ],
        "title": "InvoiceResponse"
      },
...
    }
  }
}